Efficiency gains by only registering for in-game events when required, plus a lot of cleanup
Define local functions that register the addon for in-game events (like interacting with a Merchant, or learning a spell) when the addon is shown, and un-register the addon from those in-game events when the addon is hidden.

Additionally, remove variable definitions and button creation logic that had been re-written elsewhere in the code.