Final workaround for lack of Titan Child Buttons, using Titan Labels
In short:
- for each button (aka spell reagent), work out the width (in chars) of the Icon and Text that will be each button's label
- adjust the vertical and horizontal spacing to make things look good
- track the total width required for all buttons

Then:
- work out what font the addon is using
- work out the width of a space character of that font, in pixels
- create enough space on the Titan Panel to cater for all reagents that will be shown, using the total width for buttons from earlier and the width in pixels of a single character in that font

This is needed as the dynamically created button frames float on top of the parent frame. If we didn't create enough space for the buttons in the parent frame, then they'd overlap to other Titan stuff. No bueno.

Also, use included Titan libraries instead of loading other stuff in.